@@ -1,8 +1,3 @@
-module github.com/sevlyar/go-daemon
+module github.com/gnuos/daemon
 
-go 1.13
-
-require (
-	github.com/kardianos/osext v0.0.0-20190222173326-2bc1f35cddc0
-	golang.org/x/sys v0.17.0
-)
+go 1.24.10
