@@ -1,12 +1,13 @@
 //go:build !darwin && !dragonfly && !freebsd && !linux && !netbsd && !openbsd && !plan9 && !solaris
 // +build !darwin,!dragonfly,!freebsd,!linux,!netbsd,!openbsd,!plan9,!solaris
 
+// Package daemon
 package daemon
 
-func lockFile(fd uintptr) error {
+func lockFile(_ uintptr) error {
 	return errNotSupported
 }
 
-func unlockFile(fd uintptr) error {
+func unlockFile(_ uintptr) error {
 	return errNotSupported
 }
