@@ -1,6 +1,4 @@
-//go:build go1.8
-// +build go1.8
-
+// Package daemon
 package daemon
 
 import (
